+++ final/Assets/Scripts/GameManager.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 
 public class GameManager : MonoBehaviour
 {
-    public float initiativeNumber = 0;
+    public int initiativeNumber = 0;
 
     // Start is called before the first frame update
     void Start()
